w3c-reports/12092014/gtk/runned-tests/html/semantics/forms/the-button-element/button-validation-expected.txt

button button button button button button

PASS missing type attribute 
PASS submit type attribute 
PASS reset type attribute 
PASS button type attribute 
FAIL menu type attribute assert_equals: expected "menu" but got "submit"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/the-button-element/button-validation.html:18:20
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
willValid@http://localhost:8800/html/semantics/forms/the-button-element/button-validation.html:17:9
global code@http://localhost:8800/html/semantics/forms/the-button-element/button-validation.html:27:12)
PASS invalid type attribute 
